category tables,model and controller added and category crud completed

+++ database/migrations/2024_10_06_073131_create_inventories_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->text('title');
             $table->bigInteger('product_id')->unsigned();
             $table->bigInteger('brand_id')->unsigned();
-            $table->bigInteger('category_id')->unsigned();
+            $table->integer('category_id')->unsigned();
             $table->string('sku', 200);
             $table->enum('condition', ['New', 'Used', 'Refurbished']);
             $table->text('condition_note')->nullable();
